Rework state of most Element components into properties, and rework the configuration menu
