+++ src/main/java/net/ilexiconn/llibrary/client/gui/element/CheckboxElement.java
@@ -1,23 +1,21 @@
 package net.ilexiconn.llibrary.client.gui.element;
 
 import net.ilexiconn.llibrary.LLibrary;
+import net.ilexiconn.llibrary.server.property.IBooleanProperty;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
-import java.util.function.Function;
-
 @SideOnly(Side.CLIENT)
 public class CheckboxElement<T extends IElementGUI> extends Element<T> {
-    private boolean selected;
-    private Function<CheckboxElement<T>, Boolean> function;
+    public final IBooleanProperty selected;
 
-    public CheckboxElement(T gui, float posX, float posY) {
-        this(gui, posX, posY, null);
+    public CheckboxElement(T gui, float posX, float posY, boolean selected) {
+        this(gui, posX, posY, new IBooleanProperty.WithState(selected));
     }
 
-    public CheckboxElement(T handler, float posX, float posY, Function<CheckboxElement<T>, Boolean> function) {
-        super(handler, posX, posY, 12, 12);
-        this.function = function;
+    public CheckboxElement(T gui, float posX, float posY, IBooleanProperty selected) {
+        super(gui, posX, posY, 12, 12);
+        this.selected = selected;
     }
 
     @Override
@@ -27,7 +25,7 @@ public class CheckboxElement<T extends IElementGUI> extends Element<T> {
         } else {
             this.drawRectangle(this.getPosX() + 1, this.getPosY() + 1, this.getWidth() - 1, this.getHeight() - 1, LLibrary.CONFIG.getTertiaryColor());
         }
-        if (this.selected) {
+        if (this.selected.getBoolean()) {
             this.drawRectangle(this.getPosX() + 3, this.getPosY() + 3, this.getWidth() - 5, this.getHeight() - 5, LLibrary.CONFIG.getTextColor());
         }
     }
@@ -35,11 +33,8 @@ public class CheckboxElement<T extends IElementGUI> extends Element<T> {
     @Override
     public boolean mouseClicked(float mouseX, float mouseY, int button) {
         if (button == 0 && super.isSelected(mouseX, mouseY)) {
-            this.selected = !this.selected;
-            if (this.function != null && this.function.apply(this)) {
+            if (this.selected.trySetBoolean(!this.selected.getBoolean())) {
                 this.gui.playClickSound();
-            } else {
-                this.selected = !this.selected;
             }
             return true;
         }
@@ -47,17 +42,12 @@ public class CheckboxElement<T extends IElementGUI> extends Element<T> {
     }
 
     public CheckboxElement<T> withSelection(boolean selected) {
-        this.selected = selected;
-        return this;
-    }
-
-    public CheckboxElement<T> withFunction(Function<CheckboxElement<T>, Boolean> function) {
-        this.function = function;
+        this.selected.trySetBoolean(selected);
         return this;
     }
 
     public boolean isSelected() {
-        return selected;
+        return this.selected.getBoolean();
     }
 }
