+++ src/main/java/net/ilexiconn/llibrary/client/gui/element/DropdownButtonElement.java
@@ -1,30 +1,25 @@
 package net.ilexiconn.llibrary.client.gui.element;
 
 import net.ilexiconn.llibrary.LLibrary;
+import net.ilexiconn.llibrary.server.property.IStringSelectionProperty;
 import net.minecraft.client.gui.FontRenderer;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
-import java.util.List;
-import java.util.function.Function;
-
 @SideOnly(Side.CLIENT)
 public class DropdownButtonElement<T extends IElementGUI> extends Element<IElementGUI> {
-    private final List<String> entries;
+    private IStringSelectionProperty selected;
     private String text;
     private boolean dropped;
-    private Function<DropdownButtonElement<T>, Boolean> function;
     private int dropdownWidth;
-    private String selected;
 
-    public DropdownButtonElement(T gui, String text, float posX, float posY, int width, int height, List<String> entries, Function<DropdownButtonElement<T>, Boolean> function) {
+    public DropdownButtonElement(T gui, float posX, float posY, int width, int height, IStringSelectionProperty selected) {
         super(gui, posX, posY, width, height);
-        this.text = text;
-        this.function = function;
-        this.entries = entries;
+        this.selected = selected;
+        this.text = selected.getString();
 
-        for (String entry : entries) {
-            int entryWidth = gui.getFontRenderer().getStringWidth(entry) + 5;
+        for (String value : this.selected.getValidStringValues()) {
+            int entryWidth = gui.getFontRenderer().getStringWidth(value) + 5;
             if (entryWidth > dropdownWidth) {
                 this.dropdownWidth = entryWidth;
             }
@@ -37,9 +32,9 @@ public class DropdownButtonElement<T extends IElementGUI> extends Element<IEleme
         FontRenderer fontRenderer = this.gui.getFontRenderer();
         fontRenderer.drawString(this.text, this.getPosX() + (this.getWidth() / 2) - (fontRenderer.getStringWidth(this.text) / 2), this.getPosY() + (this.getHeight() / 2) - (fontRenderer.FONT_HEIGHT / 2), LLibrary.CONFIG.getTextColor(), false);
         if (this.dropped) {
-            this.drawRectangle(this.getPosX(), this.getPosY() + this.getHeight(), this.dropdownWidth, this.entries.size() * 12, this.getColorScheme().getSecondaryColor());
+            this.drawRectangle(this.getPosX(), this.getPosY() + this.getHeight(), this.dropdownWidth, this.selected.getValidStringValues().size() * 12, this.getColorScheme().getSecondaryColor());
             float y = this.getPosY() + this.getHeight() + 2;
-            for (String entry : this.entries) {
+            for (String entry : this.selected.getValidStringValues()) {
                 if (this.isEntrySelected(mouseX, mouseY, y)) {
                     this.drawRectangle(this.getPosX(), y - 2, this.dropdownWidth, 12, this.getColorScheme().getPrimaryColor());
                 }
@@ -59,10 +54,9 @@ public class DropdownButtonElement<T extends IElementGUI> extends Element<IEleme
         } else {
             if (this.dropped) {
                 float y = this.getPosY() + this.getHeight() + 2;
-                for (String entry : this.entries) {
+                for (String value : this.selected.getValidStringValues()) {
                     if (this.isEntrySelected(mouseX, mouseY, y)) {
-                        this.selected = entry;
-                        if (this.function.apply(this)) {
+                        if (this.selected.trySetString(value)) {
                             this.gui.playClickSound();
                             break;
                         }
@@ -76,7 +70,7 @@ public class DropdownButtonElement<T extends IElementGUI> extends Element<IEleme
     }
 
     public String getSelected() {
-        return this.selected;
+        return this.selected.getString();
     }
 
     private boolean isEntrySelected(float mouseX, float mouseY, float y) {
